Backend/requirements.txt
Flask
pymongo
python-dotenv
ipfshttpclient
w3storage
gunicorn
langchain-groq
requests
eth-account
flask-cors

